refactor: extract DummyInput* into src/DummyInput.ts
Utils.ts was ~1780 lines and roughly half of its weight came from the
dummy-input focus-sentinel system: four tightly-coupled classes plus a
debug helper and a shared priority map. It is a standalone focus-sentinel
feature, not a grab-bag utility.

Moved:
- DummyInput, DummyInputProps, DummyInputFocusCallback
- DummyInputManager (+ static moveWithPhantomDummy /
  addPhantomDummyWithTarget)
- DummyInputManagerPriorities constant
- DummyInputManagerCore (internal)
- DummyInputObserver
- DummyInputWrapper (internal) and setDummyInputDebugValue (internal,
  __DEV__-gated)
- getDummyInputContainer
- Supporting internal interfaces (HTMLElementWithDummyContainer,
  HTMLElementWithDummyInputs) and the _updateDummyInputsTimeout constant

Kept in Utils.ts: WeakHTMLElement, hasSubFocusable, makeFocusIgnored —
they're imported back by DummyInput.ts. Unused imports that only served
the moved code (nativeFocus, TabsterMoveFocusEvent, SysProps,
SysDummyInputsPositions, TABSTER_DUMMY_INPUT_ATTRIBUTE_NAME, the
DummyInputObserver interface alias) are dropped from Utils.ts.

Import sites updated in 8 files (Focusable, Groupper, Modalizer, Mover,
Root, State/FocusedElement, Tabster, index). The public re-export of
getDummyInputContainer now comes directly from DummyInput.ts via
src/index.ts — Tabster.ts no longer proxies it.

Bundle size: identical to the byte on every fixture. This is a structural
change; the win is readability plus per-module visibility in future
source-map-explorer reports.
